refactor(error-handling): don't show special error text on duplication errors
the initial idea was to show a special error message that tells the user that the import currently
does not support updating. But if the user did not want to update something in the first place he
would be interested in knowing which field was causing the duplication error. That's why we now show
the original error from the API.